requirements.txt
astroid==2.2.5
blinker==1.3
ecdsa==0.13
falcon==1.4.1
Flask==0.10.1
Flask-Cors==2.0.1
Flask-JWT==0.3.2
Flask-Moment==0.4.0
Flask-Script==2.0.5
Flask-WTF==0.12
future==0.16.0
gunicorn==19.7.1
isort==4.3.17
itsdangerous==0.24
Jinja2==2.7.3
lazy-object-proxy==1.3.1
MarkupSafe==0.23
mccabe==0.6.1
num2words==0.5.6
paramiko==1.15.2
pdfkit==0.6.1
peewee==3.7.1
phonenumbers==8.10.10
pika==1.1.0
psycopg2==2.8.2
pycrypto==2.6.1
PyJWT==1.4.2
pylint==2.3.1
pymongo==3.2.2
python-dateutil==2.6.1
python-mimeparse==1.6.0
redis==2.10.6
requests==2.7.0
rosapi==0.2.5
six==1.11.0
stdlib-list==0.4.0
typed-ast==1.3.4
Unidecode==1.0.22
urllib3==1.22
uuid==1.30
vk==1.5.3
Werkzeug==0.10.4
wrapt==1.11.1
WTForms==2.2.1
